moved app state to app file and described errors
